Add Test proyect and reorder
