+++ LuigimonORM.sln
@@ -3,7 +3,9 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.1.32210.238
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "LuigimonORM", "LuigimonORM.csproj", "{3BBC4EA1-D805-4A91-AE09-5B45AB01CBA0}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "LuigimonORM", "LuigimonORM\LuigimonORM.csproj", "{3BBC4EA1-D805-4A91-AE09-5B45AB01CBA0}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Test_LuigimonORM", "Test_LuigimonORM\Test_LuigimonORM.csproj", "{242E0CAA-BC3D-4A68-BB10-636E000A5971}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -15,6 +17,10 @@ Global
 		{3BBC4EA1-D805-4A91-AE09-5B45AB01CBA0}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{3BBC4EA1-D805-4A91-AE09-5B45AB01CBA0}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{3BBC4EA1-D805-4A91-AE09-5B45AB01CBA0}.Release|Any CPU.Build.0 = Release|Any CPU
+		{242E0CAA-BC3D-4A68-BB10-636E000A5971}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{242E0CAA-BC3D-4A68-BB10-636E000A5971}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{242E0CAA-BC3D-4A68-BB10-636E000A5971}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{242E0CAA-BC3D-4A68-BB10-636E000A5971}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
