+++ LuigimonORM/ITableNameResolver.cs
