+++ Test_LuigimonORM/EntityTest.cs
@@ -1,0 +1,13 @@
+using Microsoft.VisualStudio.TestTools.UnitTesting;
+
+namespace Test_LuigimonORM
+{
+    [TestClass]
+    public class EntityTest
+    {
+        [TestMethod]
+        public void TestMethod1()
+        {
+        }
+    }
+}
